chore: add support for the uv package manager
This is mostly for convenience purposes and generating simple
dependency graphs. Note that you most likely still need to be on a
Debian-based Linux distribution for this to work. Debian trixie is
used at the time of writing.

If you stumbled across this commit for some reason and also wish to
use `uv`, you can do the following after installing `uv`[0]:

```bash
uv python install cpython-3.13.5-linux-x86_64-gnu
uv venv --python 3.15.5
source .venv/bin/activate
```

After that, a call to `uv tree` should be enough in order to generate
a dependency tree.

[0]: https://docs.astral.sh/uv/getting-started/installation/

Signed-off-by: Max R. Carrara <max@aequito.sh>